refactor(arbitrate): relocate ArbitrationCategory to un-gated module
